Implemented New Profile Database
Made necessary modifications to all databases to include a profile ID foreign key. The foreign key is used to retrieve only the data for a selected user. Methods have been updated to use the profile ID. Profile ID is passed into activities via intents.
Currently, the report activity is not working properly. It needs to be updated to use the profile ID in the retrieval of session info.